# Conflicts:
#	docs/UserGuide.md
#	src/main/java/seedu/duke/item/Item.java
#	src/main/java/seedu/duke/transaction/Transaction.java
#	src/test/java/seedu/duke/storage/TransactionStorageTest.java
#	src/test/java/seedu/duke/transaction/TransactionTest.java

+++ src/main/java/seedu/duke/transaction/TransactionList.java
@@ -3,6 +3,7 @@ package seedu.duke.transaction;
 import seedu.duke.exception.InvalidTransactionException;
 import seedu.duke.exception.TransactionNotFoundException;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -68,7 +69,7 @@ public class TransactionList {
      */
 
     //@@author jorellesee
-    public Transaction updateTransaction(String transactionId, int duration, double moneyTransacted)
+    public Transaction updateTransaction(String transactionId, int duration, BigDecimal moneyTransacted)
             throws TransactionNotFoundException, InvalidTransactionException {
         for (int i = 0; i < this.transactionList.size(); ++i) {
             Transaction tx = this.transactionList.get(i);
@@ -126,7 +127,7 @@ public class TransactionList {
     public double getTotalMoneyTransacted() {
         int totalProfit = 0;
         for (Transaction transaction : transactionList) {
-            totalProfit += transaction.getMoneyTransacted();
+            totalProfit += transaction.getMoneyTransacted().doubleValue();
         }
         return totalProfit;
     }
